Add a typeclass for finding instr rules
This will allow distributed control of determining which instrrule to
apply for a given instruction.  We don't use it yet, because it's late and
I have to run and haven't had time to apply it yet, but I'll make use of
it tomorrow in a general tactic that deals with both structured and
unstructured code.