Be more fine-grained in the while rule
We should only have to prove that transitioning preserves the
postcondition when we might actually transition.  Of course, this
requires deciding in advance whether or not we're actually going to
transition (or, at least, I don't have a good way to make it work
otherwise), so we require that the logical bit be decidable.